requirements.txt
# ── Web framework ─────────────────────────────────────────────
django>=4.2,<5.0
django-json-widget>=1.1.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.9

# ── Supabase ──────────────────────────────────────────────────
supabase>=2.4.0

# ── Scraping ──────────────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
cloudscraper>=1.2.71
playwright>=1.40.0

# ── AI / LLM ──────────────────────────────────────────────────
google-genai>=1.0.0

# ── PDF ───────────────────────────────────────────────────────
pdfplumber>=0.10.0

# ── Utils ─────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0
